+++ spec/features/projects/users_can_view_a_list_of_regional_casework_services_projects_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 RSpec.feature "Viewing regional casework services projects" do
   context "when there are no projects" do
     before do
-      user = create(:user, :caseworker)
+      user = create(:user, :team_leader)
       sign_in_with_user(user)
     end
 
@@ -22,108 +22,62 @@ RSpec.feature "Viewing regional casework services projects" do
     end
   end
 
-  context "when there are projects" do
+  context "navigation" do
     before do
       sign_in_with_user(user)
-      mock_successful_api_response_to_create_any_project
-      mock_pre_fetched_api_responses_for_any_establishment_and_trust
     end
 
-    let!(:completed_regional_project) { create(:conversion_project, urn: 126041, team: "london", completed_at: Date.yesterday) }
-    let!(:in_progress_regional_project) { create(:conversion_project, urn: 126041, team: "london", assigned_to: user) }
-
-    let!(:completed_project) { create(:conversion_project, urn: 121583, completed_at: Date.yesterday, team: "regional_casework_services", assigned_to: user) }
-    let!(:in_progress_project) { create(:conversion_project, urn: 115652, team: "regional_casework_services", assigned_to: user) }
-
-    let!(:unassigned_project) { create(:conversion_project, urn: 103835, team: "regional_casework_services", assigned_to: nil) }
-
-    context "when signed in as a Regional caseworker" do
-      let(:user) { create(:user, :caseworker) }
-
-      scenario "they can view all in progress projects" do
-        view_in_progress_projects
-      end
-
-      scenario "they can view all in progress projects" do
-        view_completed_projects
-      end
-
-      scenario "they can view unassigned projects" do
-        view_unassigned_projects
-      end
-    end
-
-    context "when signed in as a Regional caseworker team lead" do
+    context "when the user is a team_leader" do
       let(:user) { create(:user, :team_leader) }
 
-      scenario "they can view all in progress projects" do
-        view_in_progress_projects
+      scenario "they can see a link to the unassigned projects page" do
+        visit in_progress_team_projects_path
+
+        expect(page.find("nav.moj-primary-navigation")).to have_content "Unassigned"
       end
 
-      scenario "they can view all in progress projects" do
-        view_completed_projects
-      end
+      scenario "the Team projects link leads to the unassigned page" do
+        visit root_path
 
-      scenario "they can view unassigned projects" do
-        view_unassigned_projects
+        click_on "Team projects"
+
+        expect(page.find("h1")).to have_content("Unassigned")
       end
     end
 
-    context "when signed in as a Regional delivery officer" do
+    context "when the user is a regional delivery officer" do
       let(:user) { create(:user, :regional_delivery_officer) }
 
-      scenario "they can view all in progress projects" do
-        view_in_progress_projects
+      scenario "they can NOT see a link to the unassigned projects page" do
+        visit in_progress_team_projects_path
+
+        expect(page.find("nav.moj-primary-navigation")).to_not have_content "Unassigned"
       end
 
-      scenario "they can view all in progress projects" do
-        view_completed_projects
-      end
+      scenario "the Team projects link leads to the in-progress page" do
+        visit root_path
 
-      scenario "they can view unassigned projects" do
-        view_unassigned_projects
-      end
-    end
+        click_on "Team projects"
 
-    def view_in_progress_projects
-      visit in_progress_team_projects_path
-
-      expect(page).to have_content(I18n.t("project.team.in_progress.title"))
-
-      within("tbody") do
-        expect(page).to have_content(in_progress_project.urn)
-        expect(page).not_to have_content(in_progress_regional_project.urn)
-
-        expect(page).not_to have_content(completed_project.urn)
-        expect(page).not_to have_content(completed_regional_project.urn)
+        expect(page.find("h1")).to have_content("In progress")
       end
     end
 
-    def view_completed_projects
-      visit completed_team_projects_path
+    context "when the user is a caseworker" do
+      let(:user) { create(:user, :caseworker) }
 
-      expect(page).to have_content(I18n.t("project.team.completed.title"))
+      scenario "they can NOT see a link to the unassigned projects page" do
+        visit in_progress_team_projects_path
 
-      within("tbody") do
-        expect(page).to have_content(completed_project.urn)
-        expect(page).not_to have_content(completed_regional_project.urn)
-
-        expect(page).not_to have_content(in_progress_project.urn)
-        expect(page).not_to have_content(in_progress_regional_project.urn)
+        expect(page.find("nav.moj-primary-navigation")).to_not have_content "Unassigned"
       end
-    end
 
-    def view_unassigned_projects
-      visit unassigned_team_projects_path
+      scenario "the Team projects link leads to the in-progress page" do
+        visit root_path
 
-      expect(page).to have_content(I18n.t("project.team.unassigned.title"))
+        click_on "Team projects"
 
-      within("tbody") do
-        expect(page).not_to have_content(completed_project.urn)
-        expect(page).not_to have_content(completed_regional_project.urn)
-
-        expect(page).not_to have_content(in_progress_project.urn)
-        expect(page).not_to have_content(in_progress_regional_project.urn)
+        expect(page.find("h1")).to have_content("In progress")
       end
     end
   end