+++ CHANGELOG.md
@@ -20,6 +20,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Dates in csv exports are now formated as YYYY-MM-DD.
 - The application now uses the DfE Design System fonts.
 - The application now uses the DfE Design System header, footer and buttons.
+- Change the Team projects views to only show projects aligned to the user's
+  team
+- Only team leaders can see the Unassigned projects page in the Team projects
+  view
+- All users with a role and a team can now view the Team projects view
 
 ### Fixed